Keep DOM utility tests easier to extend
Queue item 111 targets the DOM utility test suite, so this extracts the shared DOM/global mock setup into a sibling helper module while preserving the existing assertions and browser-facing behavior coverage.

Constraint: Preserve dom-utils test behavior and keep the refactor limited to local test helper extraction in veryfront-code
Rejected: Split the test suite into multiple files by function | broader file movement and review scope than this queue item needs
Confidence: high
Scope-risk: narrow
Directive: Keep DOM-specific mock factories near dom-utils tests unless another routing client test needs the exact same mock layer
Tested: deno fmt --check src/routing/client/dom-utils.test.ts src/routing/client/dom-utils.test-helpers.ts; deno lint src/routing/client/dom-utils.test.ts src/routing/client/dom-utils.test-helpers.ts; deno check src/routing/client/dom-utils.test.ts src/routing/client/dom-utils.test-helpers.ts; deno test --allow-all src/routing/client/dom-utils.test.ts; git diff --check
Not-tested: Additional veryfront-code suites beyond the focused dom-utils test coverage
Not-tested: Repository pre-commit verify:quick is currently blocked by unrelated existing CLI boundary violations in cli/commands/extension/*.ts
